set types

+++ src/infinity.c
@@ -268,7 +268,7 @@ static struct PyMethodDef methods[] = {
 
 
 PyTypeObject segments_Infinity_Type = {
-	PyObject_HEAD_INIT(NULL)
+	PyObject_HEAD_INIT((long int) NULL)
 	.tp_as_number = &as_number,
 	.tp_basicsize = sizeof(segments_Infinity),
 	.tp_doc =
