+++ src/segment.c
@@ -480,7 +480,7 @@ static struct PyMethodDef methods[] = {
 
 
 PyTypeObject segments_Segment_Type = {
-	PyObject_HEAD_INIT(NULL)
+	PyObject_HEAD_INIT((long int) NULL)
 	.tp_as_number = &as_number,
 	.tp_as_sequence = &as_sequence,
 	.tp_doc =
